+++ src/Gate.IO.Api/Swap/Requests/GateSwapMarketQueryRequest.cs
@@ -1,0 +1,22 @@
+namespace Gate.IO.Api.Swap;
+
+/// <summary>
+/// Flash swap currency pair query request
+/// </summary>
+public record GateSwapMarketQueryRequest
+{
+    /// <summary>
+    /// Query by specified currency name
+    /// </summary>
+    public string Currency { get; set; }
+
+    /// <summary>
+    /// Page number
+    /// </summary>
+    public int? Page { get; set; }
+
+    /// <summary>
+    /// Maximum number of items returned
+    /// </summary>
+    public int? Limit { get; set; }
+}